Refactoring of Expressions - break large file into 3
