+++ BugReport/Query/Expressions.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Text.RegularExpressions;
 using BugReport.DataModel;
 using BugReport.Util;
 
@@ -181,817 +180,6 @@ namespace BugReport.Query
     {
         public ExpressionTooComplexException()
         {
-        }
-    }
-
-    public class ExpressionNot : Expression
-    {
-        readonly Expression _expr;
-
-        public Expression Expression
-        {
-            get => _expr;
-        }
-
-        public ExpressionNot(Expression expr)
-        {
-            _expr = expr;
-        }
-
-        public override bool Evaluate(DataModelIssue issue)
-        {
-            return !_expr.Evaluate(issue);
-        }
-
-        public override void Validate(IssueCollection collection)
-        {
-            _expr.Validate(collection);
-        }
-
-        public override string ToString()
-        {
-            if ((_expr is ExpressionAnd) || (_expr is ExpressionOr) || (_expr is ExpressionMultiRepo))
-            {
-                return $"!({_expr})";
-            }
-            return $"-{_expr}";
-        }
-
-        public override string GetGitHubQueryURL()
-        {
-            if ((_expr is ExpressionLabel) || (_expr is ExpressionMilestone) || (_expr is ExpressionAssignee))
-            {
-                return "-" + _expr.GetGitHubQueryURL();
-            }
-            if (_expr == ExpressionConstant.False)
-            {
-                return "";
-            }
-            return null;
-        }
-
-        internal override bool IsNormalized(NormalizedState minAllowedState)
-        {
-            return (minAllowedState <= NormalizedState.Not) &&_expr.IsNormalized(NormalizedState.Leaf);
-        }
-
-        protected override Expression GetSimplified()
-        {
-            if (_expr is ExpressionNot)
-            {
-                return ((ExpressionNot)_expr)._expr.Simplified;
-            }
-
-            Expression simplifiedExpr = _expr.Simplified;
-            if (simplifiedExpr is ExpressionOr)
-            {
-                return Expression.And(((ExpressionOr)simplifiedExpr).Expressions.Select(e => Expression.Not(e))).Simplified;
-            }
-            if (simplifiedExpr is ExpressionAnd)
-            {
-                return Expression.Or(((ExpressionAnd)simplifiedExpr).Expressions.Select(e => Expression.Not(e))).Simplified;
-            }
-            if (simplifiedExpr is ExpressionMultiRepo)
-            {
-                return new ExpressionMultiRepo(((ExpressionMultiRepo)simplifiedExpr).RepoExpressions.Select(repoExpr =>
-                    new RepoExpression(repoExpr.Repo, Expression.Not(repoExpr.Expr).Simplified)));
-            }
-            if (_expr is ExpressionConstant)
-            {
-                Debug.Assert((_expr == ExpressionConstant.True) || (_expr == ExpressionConstant.False));
-                return (_expr == ExpressionConstant.True) ? ExpressionConstant.False : ExpressionConstant.True;
-            }
-            return Expression.Not(simplifiedExpr);
-        }
-
-        public override bool Equals(Expression e)
-        {
-            return (e is ExpressionNot) &&
-                ((ExpressionNot)e)._expr.Equals(_expr);
-        }
-    }
-
-    public class ExpressionAnd : Expression
-    {
-        readonly IEnumerable<Expression> _expressions;
-
-        public IEnumerable<Expression> Expressions
-        {
-            get => _expressions;
-        }
-
-        public ExpressionAnd(IEnumerable<Expression> expressions)
-        {
-            _expressions = expressions;
-        }
-
-        public override bool Evaluate(DataModelIssue issue)
-        {
-            foreach (Expression expr in _expressions)
-            {
-                if (!expr.Evaluate(issue))
-                {
-                    return false;
-                }
-            }
-            return true;
-        }
-
-        public override void Validate(IssueCollection collection)
-        {
-            foreach (Expression expr in _expressions)
-            {
-                expr.Validate(collection);
-            }
-        }
-
-        public override string ToString()
-        {
-            return string.Join(" ", _expressions.Select(e => (e is ExpressionOr) ? $"({e})" : e.ToString()));
-        }
-
-        public override string GetGitHubQueryURL()
-        {
-            IEnumerable<string> subQueries = _expressions.Select(expr => expr.GetGitHubQueryURL());
-            if (subQueries.Contains(null))
-            {
-                return null;
-            }
-            return string.Join(" ", subQueries);
-        }
-
-        internal override bool IsNormalized(NormalizedState minAllowedState)
-        {
-            return (minAllowedState <= NormalizedState.And) &&
-                _expressions.Where(e => !e.IsNormalized(NormalizedState.And + 1)).None();
-        }
-
-        static readonly int const_SubExpressionsLimit = 100;
-
-        protected override Expression GetSimplified()
-        {
-            List<Expression> andExpressions = new List<Expression>();
-
-            // First flatten all ANDs
-            Queue<Expression> simplifiedExpressionsQueue = new Queue<Expression>(_expressions);
-            while (simplifiedExpressionsQueue.Count > 0)
-            {
-                Expression simplifiedExpression = simplifiedExpressionsQueue.Dequeue();
-                if (simplifiedExpression is ExpressionAnd)
-                {   // Fold all inner AND operands into this one
-                    foreach (Expression expr2 in ((ExpressionAnd)simplifiedExpression)._expressions)
-                    {
-                        simplifiedExpressionsQueue.Enqueue(expr2);
-                    }
-                }
-                else if (simplifiedExpression == ExpressionConstant.True)
-                {   // Skip 'True' expressions in AND list
-                }
-                else
-                {
-                    andExpressions.Add(simplifiedExpression);
-                }
-            }
-                
-            // Now normalize the AND expressions (can't do it earlier, because AND(x,AND(y,OR(a,b))) would start to permutate too early
-            {
-                List<Expression> andExpressionsSimplified = new List<Expression>(andExpressions.Select(e => e.Simplified));
-                andExpressions = andExpressionsSimplified;
-            }
-
-            // Handle multi-repo sub-expressions if present - bubble merge multi-repo expression up
-            IEnumerable<ExpressionMultiRepo> multiRepoExpressions = andExpressions
-                .Where(e => e is ExpressionMultiRepo)
-                .Select(e => (ExpressionMultiRepo)e);
-            if (multiRepoExpressions.Any())
-            {
-                IEnumerable<Repository> repos = multiRepoExpressions
-                    .SelectMany(e => e.RepoExpressions.Select(re => re.Repo))
-                    .Distinct();
-                IEnumerable<RepoExpression> repoExpressions = repos
-                    .Select(repo =>
-                        new RepoExpression(repo,
-                            Expression.And(andExpressions.Select(e =>
-                                    (e is ExpressionMultiRepo)
-                                        ? ((ExpressionMultiRepo)e).GetExpression(repo)
-                                        : e))
-                                // We need to normalize to bubble up all ORs on top
-                                .Simplified
-                        )
-                    );
-                Expression multiRepoExpr = new ExpressionMultiRepo(repoExpressions);
-                return multiRepoExpr;
-            }
-
-            // Handle OR sub-expressions if present - make sub-expression permutations and bubble OR up
-            IEnumerable<ExpressionOr> orSubExpressions = andExpressions
-                .Where(e => e is ExpressionOr)
-                .Select(e => (ExpressionOr)e).ToArray();
-            if (orSubExpressions.Any())
-            {
-                IEnumerable<Expression> nonOrSubExpressions = andExpressions.Where(e => !(e is ExpressionOr)).ToArray();
-                Debug.Assert(andExpressions.Count() == orSubExpressions.Count() + nonOrSubExpressions.Count());
-
-                // Check for large expansions
-                int combinationsCount = 1;
-                foreach (ExpressionOr orSubExpression in orSubExpressions)
-                {
-                    combinationsCount *= orSubExpression.Expressions.Count();
-                }
-                if (combinationsCount < const_SubExpressionsLimit)
-                {
-                    List<Expression> orExpressions = new List<Expression>();
-
-                    // (A || B) && Z ---> (A && Z) || (B && Z)
-                    // (A || B) && (C || D) && Z ---> (A && C && Z) || (A && D && Z) || (B && C && Z) || (B && D && Z)
-                    foreach (IEnumerable<Expression> orSubExpressionsAndPermutation in GeneratePermutations(orSubExpressions))
-                    {
-                        orExpressions.Add(Expression.And(orSubExpressionsAndPermutation.Concat(nonOrSubExpressions)).Simplified);
-                    }
-                    Expression orExpression = Expression.Or(orExpressions);
-                    return orExpression;
-                }
-            }
-
-            // Simplify the expression
-            if (andExpressions.Contains(ExpressionConstant.False))
-            {
-                return ExpressionConstant.False;
-            }
-            RemoveDuplicates(andExpressions);
-            if (ContainsNegatedExpressions(andExpressions))
-            {
-                return ExpressionConstant.False;
-            }
-
-            if (andExpressions.Count == 1)
-            {
-                return andExpressions[0];
-            }
-
-            // AND is on the top
-            Expression andExpression = Expression.And(andExpressions);
-            return andExpression;
-        }
-
-        private IEnumerable<IEnumerable<Expression>> GeneratePermutations(IEnumerable<ExpressionOr> orExpressions)
-        {
-            return GeneratePermutations(orExpressions, orExpressions.Count());
-        }
-        private IEnumerable<IEnumerable<Expression>> GeneratePermutations(
-            IEnumerable<ExpressionOr> orExpressions, 
-            int orExpressionsCount)
-        {
-            foreach (ExpressionOr orExpression in orExpressions)
-            {
-                foreach (Expression expr in orExpression.Expressions)
-                {
-                    if (orExpressionsCount == 1)
-                    {
-                        yield return expr.ToEnumerable();
-                    }
-                    else
-                    {
-                        foreach (IEnumerable<Expression> tailPermutation in GeneratePermutations(orExpressions.Skip(1)))
-                        {
-                            yield return expr.ToEnumerable().Concat(tailPermutation);
-                        }
-                    }
-                }
-            }
-        }
-
-        public override bool Equals(Expression e)
-        {
-            return (e is ExpressionAnd) ? 
-                Equals(_expressions, ((ExpressionAnd)e).Expressions) : 
-                false;
-        }
-    }
-
-    public class ExpressionOr : Expression
-    {
-        readonly IEnumerable<Expression> _expressions;
-
-        public IEnumerable<Expression> Expressions
-        {
-            get => _expressions;
-        }
-
-        public ExpressionOr(IEnumerable<Expression> expressions)
-        {
-            _expressions = expressions;
-        }
-
-        public override bool Evaluate(DataModelIssue issue)
-        {
-            foreach (Expression expr in _expressions)
-            {
-                if (expr.Evaluate(issue))
-                {
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        public override void Validate(IssueCollection collection)
-        {
-            foreach (Expression expr in _expressions)
-            {
-                expr.Validate(collection);
-            }
-        }
-        public override string ToString()
-        {
-            return string.Join(" OR ", _expressions.Select(e => (e is ExpressionAnd) ? $"({e})" : e.ToString()));
-        }
-
-        public override string GetGitHubQueryURL()
-        {
-            return null;
-        }
-
-        internal override bool IsNormalized(NormalizedState minAllowedState)
-        {
-            return (minAllowedState <= NormalizedState.Or) &&
-                _expressions.Where(e => !e.IsNormalized(NormalizedState.Or + 1)).None();
-        }
-
-        protected override Expression GetSimplified()
-        {
-            List<Expression> orExpressions = new List<Expression>();
-
-            Queue<Expression> simplifiedExpressionsQueue = new Queue<Expression>(_expressions.Select(e => e.Simplified));
-            while (simplifiedExpressionsQueue.Count > 0)
-            {
-                Expression simplifiedExpression = simplifiedExpressionsQueue.Dequeue();
-                if (simplifiedExpression is ExpressionOr)
-                {   // Fold all inner OR operands into this one
-                    foreach (Expression expr2 in ((ExpressionOr)simplifiedExpression)._expressions)
-                    {
-                        simplifiedExpressionsQueue.Enqueue(expr2);
-                    }
-                }
-                else if (simplifiedExpression == ExpressionConstant.False)
-                {   // Skip 'False' expression in OR list
-                }
-                else
-                {
-                    orExpressions.Add(simplifiedExpression);
-                }
-            }
-
-            IEnumerable<ExpressionMultiRepo> multiRepoExpressions = orExpressions
-                .Where(e => e is ExpressionMultiRepo)
-                .Select(e => (ExpressionMultiRepo)e);
-            if (multiRepoExpressions.None())
-            {
-                if (orExpressions.Contains(ExpressionConstant.True))
-                {
-                    return ExpressionConstant.True;
-                }
-                RemoveDuplicates(orExpressions);
-                if (ContainsNegatedExpressions(orExpressions))
-                {
-                    return ExpressionConstant.True;
-                }
-
-                if (orExpressions.Count == 1)
-                {
-                    return orExpressions[0];
-                }
-
-                Expression orExpr = Expression.Or(orExpressions.Distinct());
-                return orExpr;
-            }
-
-            IEnumerable<Repository> repos = multiRepoExpressions
-                .SelectMany(e => e.RepoExpressions.Select(re => re.Repo))
-                .Distinct();
-            IEnumerable<RepoExpression> repoExpressions = repos
-                .Select(repo =>
-                    new RepoExpression(repo,
-                        Expression.Or(orExpressions.Select(e =>
-                            (e is ExpressionMultiRepo)
-                                ? ((ExpressionMultiRepo)e).GetExpression(repo)
-                                : e)).Simplified
-                    )
-                );
-            Expression expr = new ExpressionMultiRepo(repoExpressions);
-            return expr;
-        }
-
-        public override bool Equals(Expression e)
-        {
-            return (e is ExpressionOr) ? 
-                Equals(_expressions, ((ExpressionOr)e).Expressions) : 
-                false;
-        }
-    }
-
-    public class ExpressionLabel : Expression
-    {
-        readonly string _labelName;
-
-        public ExpressionLabel(string labelName)
-        {
-            _labelName = labelName;
-        }
-
-        public override bool Evaluate(DataModelIssue issue)
-        {
-            return issue.HasLabel(_labelName);
-        }
-
-        public override void Validate(IssueCollection collection)
-        {
-            if (!collection.HasLabel(_labelName))
-            {
-                Console.WriteLine($"WARNING: Label does not exist: {_labelName}");
-            }
-        }
-
-        public override string ToString()
-        {
-            return GetGitHubQueryURL();
-        }
-
-        public override string GetGitHubQueryURL()
-        {
-            if (_labelName.Contains(' '))
-            {
-                return $"label:\"{_labelName}\"";
-            }
-            return "label:" + _labelName;
-        }
-
-        internal override bool IsNormalized(NormalizedState minAllowedState)
-        {
-            return true;
-        }
-
-        protected override Expression GetSimplified()
-        {
-            return this;
-        }
-
-        public override bool Equals(Expression e)
-        {
-            return (e is ExpressionLabel) ? 
-                _labelName == ((ExpressionLabel)e)._labelName : 
-                false;
-        }
-    }
-
-    public class ExpressionLabelPattern : Expression
-    {
-        readonly string _labelPattern;
-        readonly Regex _labelRegex;
-
-        public ExpressionLabelPattern(string labelPattern)
-        {
-            _labelPattern = labelPattern;
-            _labelRegex = new Regex("^" + labelPattern + "$");
-        }
-
-        public override bool Evaluate(DataModelIssue issue)
-        {
-            return issue.Labels.Where(label => _labelRegex.IsMatch(label.Name)).Any();
-        }
-
-        public override void Validate(IssueCollection collection)
-        {
-            if (collection.Labels.Where(label => _labelRegex.IsMatch(label.Name)).None())
-            {
-                Console.WriteLine($"WARNING: Label pattern does not match any label: {_labelRegex.ToString()}");
-            }
-        }
-
-        public override string ToString()
-        {
-            if (_labelPattern.Contains(' '))
-            {
-                return $"label:\"{_labelPattern}\"";
-            }
-            return "label:" + _labelPattern;
-        }
-
-        public override string GetGitHubQueryURL()
-        {
-            return null;
-        }
-
-        internal override bool IsNormalized(NormalizedState minAllowedState)
-        {
-            return true;
-        }
-
-        protected override Expression GetSimplified()
-        {
-            return this;
-        }
-
-        public override bool Equals(Expression e)
-        {
-            return (e is ExpressionLabelPattern) ?
-                _labelPattern == ((ExpressionLabelPattern)e)._labelPattern :
-                false;
-        }
-    }
-
-    public class ExpressionIsIssue : Expression
-    {
-        readonly bool _isIssue;
-
-        public ExpressionIsIssue(bool isIssue)
-        {
-            _isIssue = isIssue;
-        }
-
-        public override bool Evaluate(DataModelIssue issue)
-        {
-            return issue.IsIssueOrComment == _isIssue;
-        }
-
-        public override void Validate(IssueCollection collection)
-        {
-        }
-
-        public override string ToString()
-        {
-            return GetGitHubQueryURL();
-        }
-
-        public override string GetGitHubQueryURL()
-        {
-            return _isIssue ? "is:issue" : "is:pr";
-        }
-
-        internal override bool IsNormalized(NormalizedState minAllowedState)
-        {
-            return true;
-        }
-
-        protected override Expression GetSimplified()
-        {
-            return this;
-        }
-
-        public override bool Equals(Expression e)
-        {
-            return (e is ExpressionIsIssue) ?
-                _isIssue == ((ExpressionIsIssue)e)._isIssue :
-                false;
-        }
-    }
-
-    public class ExpressionIsOpen : Expression
-    {
-        readonly bool _isOpen;
-
-        public ExpressionIsOpen(bool isOpen)
-        {
-            _isOpen = isOpen;
-        }
-        public override bool Evaluate(DataModelIssue issue)
-        {
-            return issue.IsOpen == _isOpen;
-        }
-        public override void Validate(IssueCollection collection)
-        {
-        }
-        public override string ToString()
-        {
-            return GetGitHubQueryURL();
-        }
-
-        public override string GetGitHubQueryURL()
-        {
-            return _isOpen ? "is:open" : "is:closed";
-        }
-
-        internal override bool IsNormalized(NormalizedState minAllowedState)
-        {
-            return true;
-
-        }
-        protected override Expression GetSimplified()
-        {
-            return this;
-        }
-
-        public override bool Equals(Expression e)
-        {
-            return (e is ExpressionIsOpen) ?
-                _isOpen == ((ExpressionIsOpen)e)._isOpen :
-                false;
-        }
-    }
-
-    public class ExpressionMilestone : Expression
-    {
-        readonly string _milestoneName;
-
-        public ExpressionMilestone(string milestoneName)
-        {
-            _milestoneName = milestoneName;
-        }
-
-        public override bool Evaluate(DataModelIssue issue)
-        {
-            return issue.IsMilestone(_milestoneName);
-        }
-
-        public override void Validate(IssueCollection collection)
-        {
-            if (!collection.HasMilestone(_milestoneName))
-            {
-                Console.WriteLine($"WARNING: Milestone does not exist: {_milestoneName}");
-            }
-        }
-
-        public override string ToString()
-        {
-            return GetGitHubQueryURL();
-        }
-
-        public override string GetGitHubQueryURL()
-        {
-            if (_milestoneName == null)
-            {
-                return "no:milestone";
-            }
-            return "milestone:" + _milestoneName;
-        }
-
-        internal override bool IsNormalized(NormalizedState minAllowedState)
-        {
-            return true;
-        }
-
-        protected override Expression GetSimplified()
-        {
-            return this;
-        }
-
-        public override bool Equals(Expression e)
-        {
-            return (e is ExpressionMilestone) ?
-                _milestoneName == ((ExpressionMilestone)e)._milestoneName :
-                false;
-        }
-    }
-
-    public class ExpressionMilestonePattern : Expression
-    {
-        readonly string _milestonePattern;
-        readonly Regex _milestoneRegex;
-
-        public ExpressionMilestonePattern(string milestonePattern)
-        {
-            _milestonePattern = milestonePattern;
-            _milestoneRegex = new Regex("^" + milestonePattern + "$");
-        }
-
-        public override bool Evaluate(DataModelIssue issue)
-        {
-            return ((issue.Milestone != null) && 
-                _milestoneRegex.IsMatch(issue.Milestone.Title));
-        }
-
-        public override void Validate(IssueCollection collection)
-        {
-            // TODO: We could enumerate all Milestones on collection - very low pri
-        }
-
-        public override string ToString()
-        {
-            if (_milestonePattern.Contains(' '))
-            {
-                return $"milestone:\"{_milestonePattern}\"";
-            }
-            return "milestone:" + _milestonePattern;
-        }
-
-        public override string GetGitHubQueryURL()
-        {
-            return null;
-        }
-
-        internal override bool IsNormalized(NormalizedState minAllowedState)
-        {
-            return true;
-        }
-
-        protected override Expression GetSimplified()
-        {
-            return this;
-        }
-
-        public override bool Equals(Expression e)
-        {
-            return (e is ExpressionMilestonePattern) ?
-                _milestonePattern == ((ExpressionMilestonePattern)e)._milestonePattern :
-                false;
-        }
-    }
-
-    public class ExpressionAssignee : Expression
-    {
-        readonly string _assigneeName;
-
-        public ExpressionAssignee(string assigneeName)
-        {
-            _assigneeName = assigneeName;
-        }
-
-        public override bool Evaluate(DataModelIssue issue)
-        {
-            return issue.HasAssignee(_assigneeName);
-        }
-
-        public override void Validate(IssueCollection collection)
-        {
-            if (!collection.HasUser(_assigneeName))
-            {
-                Console.WriteLine($"WARNING: Assignee does not exist: {_assigneeName}");
-            }
-        }
-
-        public override string ToString()
-        {
-            return GetGitHubQueryURL();
-        }
-
-        public override string GetGitHubQueryURL()
-        {
-            return "assginee:" + _assigneeName;
-        }
-
-        internal override bool IsNormalized(NormalizedState minAllowedState)
-        {
-            return true;
-        }
-
-        protected override Expression GetSimplified()
-        {
-            return this;
-        }
-
-        public override bool Equals(Expression e)
-        {
-            return (e is ExpressionAssignee) ?
-                _assigneeName == ((ExpressionAssignee)e)._assigneeName :
-                false;
-        }
-    }
-
-    public class ExpressionConstant : Expression
-    {
-        readonly bool _value;
-
-        public static readonly ExpressionConstant True = new ExpressionConstant(true);
-        public static readonly ExpressionConstant False = new ExpressionConstant(false);
-
-        private ExpressionConstant(bool value)
-        {
-            _value = value;
-        }
-
-        public override bool Evaluate(DataModelIssue issue)
-        {
-            return _value;
-        }
-
-        public override void Validate(IssueCollection collection)
-        {
-        }
-
-        public override string ToString()
-        {
-            return _value.ToString();
-        }
-
-        public override string GetGitHubQueryURL()
-        {
-            return null;
-        }
-
-        internal override bool IsNormalized(NormalizedState minAllowedState)
-        {
-            return true;
-        }
-
-        protected override Expression GetSimplified()
-        {
-            return this;
-        }
-
-        public override bool Equals(Expression e)
-        {
-            // Only 2 static instances exist (True & False), so it is sufficient to compare object references
-            return (this == e);
         }
     }
 }
